\name{readmat}
\alias{readmat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
read matirx from the file
}
\description{
\code{readmat} is a wrapper function of \code{\link{scan}}
}
\usage{
readmat(file, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{
the name of a file to read data values from
}
\item{...}{
Further arguments to be passed to \code{scan}
}
}
\seealso{
\code{\link{scan}}
}
\examples{
##
}
